refactor config stuff into its own package
