+++ pathplannerlib/src/main/java/com/pathplanner/lib/config/RobotConfig.java
@@ -1,4 +1,4 @@
-package com.pathplanner.lib.trajectory.config;
+package com.pathplanner.lib.config;
 
 import com.pathplanner.lib.trajectory.MotorTorqueCurve;
 import edu.wpi.first.math.geometry.Translation2d;
